combined Snow and Snowlogic.vue
